Convert the legacy lxc bridge to networkd
This change implements the lxcbr0 interface in systemd-networkd. This is
being done to simplify how this bridge is setup and to unify it's
management across distros. The change removes all of the old dnsmasq
process and cleans up variables, templates, and files that are no longer
needed.

New tests have been added to create a container and ping an external
address in an effort to ensure the container build and host setup
process works as expected.

Change-Id: Ib2b9062b8290c9b3bccb1ae64806f86c818ad8b1
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>